# Conflicts:
#	PaceMaker/settings.py
#	PaceMaker/urls.py

+++ graduate/views.py
@@ -17,7 +17,7 @@ class GraduateLV(LoginRequiredMixin, ListView):
         scorelist = scorelist.values_list('score', flat=True)
         print(scorelist)
         for i in scorelist:
-            sum = sum + int(i)
+            sum = sum + float(i)
 
 
         return sum
@@ -28,7 +28,7 @@ class GraduateLV(LoginRequiredMixin, ListView):
         design_gradelist = StudentGrade.objects.filter(hukbun=s).values_list('grade_design', flat=True)
         sum = 0.0
         for i in design_gradelist:
-            sum = sum + int(i)
+            sum = sum + float(i)
 
 
         return sum
@@ -39,7 +39,7 @@ class GraduateLV(LoginRequiredMixin, ListView):
         sum = 0
         scorelist = StudentGrade.objects.filter(hukbun=s).filter(Q(eisu='컴과') | Q(eisu='전필')).filter(valid = '유효').values_list('score', flat=True)
         for i in scorelist:
-            sum = sum + int(i)
+            sum = sum + float(i)
 
 
         return sum
@@ -50,7 +50,7 @@ class GraduateLV(LoginRequiredMixin, ListView):
         sum = 0
         scorelist = StudentGrade.objects.filter(hukbun=s).filter(Q(eisu='M자') | Q(eisu='필수') | Q(eisu='역철') | Q(eisu='경사') | Q(eisu='체기') | Q(eisu='사회') |Q(eisu='과기')|Q(eisu='자협') | Q(eisu='미래') | Q(eisu='직필') | Q(eisu='문예') | Q(eisu='언문')).filter(valid = '유효').values_list('score', flat=True)
         for i in scorelist:
-            sum = sum + int(i)
+            sum = sum + float(i)
 
 
         return sum
@@ -61,7 +61,7 @@ class GraduateLV(LoginRequiredMixin, ListView):
         sum = 0
         scorelist = StudentGrade.objects.filter(hukbun=s).filter(eisu='M자').values_list('score', flat=True)
         for i in scorelist:
-            sum = sum + int(i)
+            sum = sum + float(i)
 
 
         return sum
@@ -72,7 +72,7 @@ class GraduateLV(LoginRequiredMixin, ListView):
         sum = 0
         scorelist = StudentGrade.objects.filter(hukbun=s).filter(Q(subject='창의적문제해결전략')|Q(subject='특허와기술개발')|Q(subject='공학윤리')|Q(subject='인간심리의이해')).values_list('score', flat=True)
         for i in scorelist:
-            sum = sum + int(i)
+            sum = sum + float(i)
 
 
         return sum
